Fix off_t printf warning

+++ src/mp3fs.c
@@ -172,7 +172,8 @@ static int mp3fs_readdir(const char *path, void *buf, fuse_fill_dir_t filler,
     while ((de = readdir(dp))) {
         struct stat st;
 
-        snprintf(origfile, strlen(origpath) + NAME_MAX + 2, "%s/%s", origpath, de->d_name);
+        snprintf(origfile, strlen(origpath) + NAME_MAX + 2, "%s/%s", origpath,
+                 de->d_name);
 
         if (lstat(origfile, &st) == -1) {
             goto stat_fail;
@@ -301,7 +302,7 @@ static int mp3fs_read(const char *path, char *buf, size_t size, off_t offset,
     int read = 0;
     struct transcoder* trans;
 
-    mp3fs_debug("read %s: %zu bytes from %jd", path, size, offset);
+    mp3fs_debug("read %s: %zu bytes from %jd", path, size, (intmax_t)offset);
 
     errno = 0;
